fix 🐞: add option for skipping static calls
Static calls are not allowed to change the state of a contract, so there might be cases where the caller doesn't want to call a function with `staticcall` first before the real call e.g when calling `payable` functions.

This comit adds the option to skip the static call entirely.

This commit also makes the `call` method of `devRegInterface` take an object as argument so to allow future updates to the function signature