improvement 🛠: make network name, contract address, and network ID highly configurable
This commit makes it possible to specify the netwok id of the network we deployed our contract to, and the contract address.

Once the network id is specified, the entire app will automatically be made to work and connect to the network which has the network id and also the contract at the address.

So if we want our app to work with contract deployed to e.g ropsten, we just specify the contract address and network id